requirements.txt
discord.py==2.4.0
protobuf==5.28.3
PyPDF2==3.0.1
python-dotenv==1.0.1
slack_bolt==1.21.2
google-generativeai
redis